Split cmdadapter black-box tests
